Update README.md
The basic example as written produces an error with recently installed versions of python and pytorch (python 3.7.1 and pytorch 1.0.1), because the numpy *_y arrays are created as type int32 which get converted to torch IntTensors, while the cross-entropy loss of pytorch expects LongTensor. This fixes the error by creating the numpy arrays with type int64. See https://pste.eu/p/2N5f.html.